fetching the data from the firebase.

+++ lib/models/grocery_item.dart
@@ -4,7 +4,6 @@ class GroceryItem{
   const GroceryItem({required this.id,required this.name,required this.quantity,required this.category});
   final String id;
   final String name;
-  final double quantity;
+  final int  quantity;
   final Category category;
-
 }
